+++ packages/clients/meldemichel/src/plugins/Gfi/Feature.vue
@@ -1,17 +1,8 @@
 <template>
   <v-card>
-    <v-icon
-      v-if="hasWindowSize && hasSmallWidth"
-      class="meldemichel-gfi-grip-icon"
-    >
-      fa-grip-lines
-    </v-icon>
-    <v-card-actions>
-      <!-- TODO implement when vector clusters are done
-      <v-btn>Prev</v-btn>
-      <v-btn>Next</v-btn>
-      -->
-      <v-spacer></v-spacer>
+    <v-card-actions v-if="!hasWindowSize || !hasSmallWidth">
+      <ActionButtons />
+      <v-spacer />
       <v-btn
         icon
         small
@@ -56,19 +47,17 @@
 </template>
 
 <script lang="ts">
-import Vue, { PropType } from 'vue'
+import Vue from 'vue'
+import { mapActions, mapGetters, mapMutations } from 'vuex'
 import { GeoJsonProperties } from 'geojson'
-import { mapActions, mapMutations, mapGetters } from 'vuex'
-
-type GfiIndexStep = -1 | 1
+import ActionButtons from './ActionButtons.vue'
 
 export default Vue.extend({
   name: 'MeldemichelGfiFeature',
+  components: {
+    ActionButtons,
+  },
   props: {
-    currentProperties: {
-      type: Object as PropType<GeoJsonProperties>,
-      required: true,
-    },
     clientWidth: {
       type: Number,
       required: true,
@@ -77,10 +66,6 @@ export default Vue.extend({
       type: String,
       default: '',
     },
-    showSwitchButtons: {
-      type: Boolean,
-      default: false,
-    },
   },
   data: () => ({
     infoFields: ['skat', 'start', 'statu'],
@@ -88,15 +73,32 @@ export default Vue.extend({
   computed: {
     ...mapGetters(['hasSmallWidth', 'hasWindowSize']),
     ...mapGetters('plugin/gfi', [
-      'windowFeatures',
+      'imageLoaded',
       'visibleWindowFeatureIndex',
+      'windowFeatures',
     ]),
     displayImage(): boolean {
       return this.currentProperties.pic
     },
+    currentProperties(): GeoJsonProperties {
+      return { ...this.windowFeatures[this.visibleWindowFeatureIndex] }
+    },
+  },
+  mounted() {
+    this.setActionButton({
+      component: ActionButtons,
+      props: {},
+    })
+  },
+  beforeDestroy() {
+    this.setActionButton(null)
   },
   methods: {
-    ...mapMutations('plugin/gfi', ['setVisibleWindowFeatureIndex']),
+    ...mapMutations('plugin/gfi', [
+      'setActionButton',
+      'setImageLoaded',
+      'setVisibleWindowFeatureIndex',
+    ]),
     ...mapActions('plugin/gfi', ['close']),
     formatProperty(type: string, value: string): string {
       if (!value) {
@@ -116,24 +118,10 @@ export default Vue.extend({
       }
       return value
     },
-    resize(): void {
-      window.dispatchEvent(new Event('resize'))
-    },
-    /** switch to next or previous feature */
-    switchFeature(by: GfiIndexStep): void {
-      const {
-        visibleWindowFeatureIndex,
-        windowFeatures,
-        setVisibleWindowFeatureIndex,
-      } = this
-      const maxIndex = windowFeatures.length - 1
-      const nextIndex = visibleWindowFeatureIndex + by
-      if (nextIndex < 0) {
-        setVisibleWindowFeatureIndex(windowFeatures.length - 1)
-      } else if (nextIndex > maxIndex) {
-        setVisibleWindowFeatureIndex(0)
-      } else {
-        setVisibleWindowFeatureIndex(nextIndex)
+    resize(e: Event) {
+      if ((e.currentTarget as HTMLImageElement).complete && !this.imageLoaded) {
+        this.setImageLoaded(true)
+        window.dispatchEvent(new Event('resize'))
       }
     },
   },
@@ -143,11 +131,6 @@ export default Vue.extend({
 <style lang="scss" scoped>
 * {
   color: #003064 !important;
-}
-
-.meldemichel-gfi-grip-icon {
-  left: 50%;
-  transition: translateX(-50%);
 }
 
 .meldemichel-fat-cell {