+++ packages/plugins/IconMenu/src/store/index.ts
@@ -1,3 +1,4 @@
+import { t } from 'i18next'
 import {
   generateSimpleGetters,
   generateSimpleMutations,
@@ -14,7 +15,7 @@ const storeModule: PolarModule<IconMenuState, IconMenuState> = {
   namespaced: true,
   state: getInitialState(),
   actions: {
-    setupModule({ commit, rootGetters }): void {
+    setupModule({ commit, dispatch, rootGetters }): void {
       const menus = rootGetters.configuration?.iconMenu?.menus || []
       const initializedMenus = menus
         .filter(({ id }) => {
@@ -44,9 +45,35 @@ const storeModule: PolarModule<IconMenuState, IconMenuState> = {
 
       const initiallyOpen =
         rootGetters.configuration?.iconMenu?.initiallyOpen || ''
-      const index = initializedMenus.findIndex(({ id }) => id === initiallyOpen)
+
+      if (!rootGetters.hasSmallWidth && !rootGetters.hasSmallHeight) {
+        dispatch('openMenuById', initiallyOpen)
+      }
+    },
+    openMenuById({ commit, getters, dispatch }, openId) {
+      const index = getters.menus.findIndex(({ id }) => id === openId)
+
       if (index !== -1) {
         commit('setOpen', index)
+        dispatch('openInMoveHandle', index)
+      }
+    },
+    openInMoveHandle({ commit, getters, rootGetters }, index: number) {
+      if (rootGetters.hasWindowSize && rootGetters.hasSmallWidth) {
+        // TODO: Add actionButton to moveHandle for vector clusters including a previous and next button. Should be added to menus
+        const { hint, id, plugin } = getters.menus[index]
+        commit(
+          'setMoveHandle',
+          {
+            closeLabel: t('plugins.iconMenu.mobileCloseButton', {
+              plugin: hint || `common:plugins.iconMenu.hints.${id}`,
+            }),
+            closeFunction: () => commit('setOpen', null),
+            component: plugin,
+            plugin: 'iconMenu',
+          },
+          { root: true }
+        )
       }
     },
   },