+++ packages/components/CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## 2.0.0
+
+- Breaking: MoveHandle can now only be rendered once which is done in `@polar/core`. Please refer to the documentation of `@polar/core` on how to use this component.
+- Feature: MoveHandle now remembers the previous position and sets it to the next content, if it doesn't exceed the maximum height.
+
 ## 1.0.0
 
 Initial release.