+++ packages/core/CHANGELOG.md
@@ -1,11 +1,16 @@
 # CHANGELOG
 
-## unpublished
+## 1.2.1
 
-- Feature: Add hovered and selected feature to vuex store. This is an optional functionality that has to be explicitly enabled and works with the `@masterportal/masterportalapi` default marker design. See configuration parameter `extendedMasterportalapiMarkers`.
+- Fix dependency `@polar/components` version.
+
+## 1.2.0
+
+- Feature: Add hovered and selected features to vuex store that support clustering. This is an optional functionality that has to be explicitly enabled and works with the `@masterportal/masterportalapi` default marker design. See configuration parameter `extendedMasterportalapiMarkers`.
 - Feature: Add zoomLevel as plugin-agnostic map information to store.
 - Feature: Change the `background-color` of all `v-tooltip`s to `#595959` and its `border` to `#fff` to be more visible. It now always has a contrast of 7, which is quite enough for AAA of WCAG.
 - Feature: Add new state variable `hasSmallDisplay` which is updated on `resize` of the `window`.
+- Feature: Add possibility to add content of plugins to the now singleton MoveHandle.
 - Chore: Add README information about listening to map client state and getters.
 
 ## 1.1.0