+++ packages/lib/createTestMountParameters/CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGELOG
 
-## unpublished
+## 1.1.0
 
 - Feature: Extend mock state to match current core state type.