+++ packages/plugins/Fullscreen/CHANGELOG.md
@@ -1,9 +1,10 @@
 # CHANGELOG
 
-## unpublished
+## 1.1.0
 
 - Feature: Hide obstructive tooltip on small devices.
 - Fix: Documentation error regarding plugin state.
+- Fix: Since iPhones do not support fullscreen mode, the fullscreen button is hidden on such devices. See [MDN](https://developer.mozilla.org/en-US/docs/Web/API/Fullscreen_API#browser_compatibility) for more information.
 
 ## 1.0.0