+++ packages/plugins/Toast/CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## 1.0.1
+
+- Fix: Add missing repository information in package.json.
+
 ## 1.0.0
 
 Initial release.